moved notification variable to navbar and added notification icon functionality
