+++ server/server.js
@@ -29,7 +29,7 @@ require("./routes/notification.route")(app);
 const NotificationController = require("./controllers/notification.controller");
 // Sends expiration date notifications every day at 7 AM PST.
 // Adjust frequency/time based on timezones if necessary.
-cron.schedule("23 18 * * *", () => {
+cron.schedule("48 14 * * *", () => {
     NotificationController.pushNotificationsToArray();
     io.emit("new-notification", { message: "New expiration notifications received!" });
 });
